Move command to `maintenance` group
